requirements.txt
aniso8601==8.0.0
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cryptography==2.9.2
flake8==3.8.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Executor==0.9.3
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
flask-restx==0.2.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pika==1.1.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.4.2
pylint==2.5.3
pyrsistent==0.16.0
python-dateutil==2.8.1
python-json-logger==0.1.11
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
toml==0.10.1
transitions==0.8.3
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
minio==7.0.3
xmltodict==0.12.0
flask-sqlalchemy==2.4.4
psycopg2==2.9.1
